# app/openapi.yaml
openapi: 3.1.0
info:
  title: Quality Books API
  version: 1.0.0
  description: Quality Books, the course's demo shop. Topic 12 feeds this spec to an AI model to propose test cases.
servers:
  - url: http://localhost:3210
paths:
  /api/books:
    get:
      summary: List or search books
      parameters:
        - name: q
          in: query
          required: false
          description: Case-insensitive match on title, author or tag.
          schema: { type: string }
      responses:
        '200':
          description: Matching books
          content:
            application/json:
              schema:
                type: object
                properties:
                  books: { type: array, items: { $ref: '#/components/schemas/Book' } }
  /api/books/{id}:
    get:
      summary: Get one book
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: The book, content: { application/json: { schema: { $ref: '#/components/schemas/Book' } } } }
        '404': { description: Unknown id, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/cart/price:
    post:
      summary: Price a cart
      description: |
        Shipping is 4.90 EUR, free when the subtotal is 50 EUR or more.
        Quantity must be an integer from 1 to 10 and not exceed stock.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [bookId, quantity]
                    properties:
                      bookId: { type: integer }
                      quantity: { type: integer, minimum: 1, maximum: 10 }
      responses:
        '200':
          description: Priced cart
          content:
            application/json:
              schema:
                type: object
                properties:
                  lines: { type: array, items: { type: object } }
                  subtotal: { type: number }
                  shipping: { type: number }
                  total: { type: number }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/assistant:
    post:
      summary: Ask the AI support assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                question: { type: string, maxLength: 1000 }
      responses:
        '200':
          description: The answer
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer: { type: string }
                  mode: { type: string, enum: [mock, buggy, claude] }
components:
  schemas:
    Book:
      type: object
      required: [id, title, author, price, stock, tags]
      properties:
        id: { type: integer }
        title: { type: string }
        author: { type: string }
        price: { type: number }
        stock: { type: integer }
        tags: { type: array, items: { type: string } }
    Error:
      type: object
      properties:
        error: { type: string }